requirements.txt
fastapi
uvicorn[standard]
pydantic
spacy
python-dotenv
python-dateutil
